no functional change; move functions to trait for consumption in a later commit
