+++ src/main/scala/org/broadinstitute/dsde/firecloud/service/TrialServiceSupport.scala
@@ -7,19 +7,28 @@ import java.util.Date
 
 import com.google.api.services.sheets.v4.model.{SpreadsheetProperties, ValueRange}
 import com.typesafe.scalalogging.LazyLogging
-import org.broadinstitute.dsde.firecloud.FireCloudException
-import org.broadinstitute.dsde.firecloud.dataaccess.{ThurloeDAO, TrialDAO}
+import org.broadinstitute.dsde.firecloud.{FireCloudException, FireCloudExceptionWithErrorReport}
+import org.broadinstitute.dsde.firecloud.dataaccess.{SamDAO, ThurloeDAO, TrialDAO}
+import org.broadinstitute.dsde.firecloud.model.Trial.StatusUpdate.Attempt
 import org.broadinstitute.dsde.firecloud.model.Trial.TrialStates.{Disabled, Enabled, TrialState}
-import org.broadinstitute.dsde.firecloud.model.Trial.{SpreadsheetResponse, TrialProject, UserTrialStatus}
+import org.broadinstitute.dsde.firecloud.model.Trial.{SpreadsheetResponse, StatusUpdate, TrialProject, UserTrialStatus}
 import org.broadinstitute.dsde.firecloud.model.{FireCloudKeyValue, Profile, ProfileWrapper, UserInfo, WorkbenchUserInfo}
+import org.broadinstitute.dsde.firecloud.service.PerRequest.{PerRequestMessage, RequestComplete}
+import org.broadinstitute.dsde.rawls.model.{RawlsBillingProjectName, RawlsUserEmail}
+import spray.http.StatusCodes
+import spray.httpx.SprayJsonSupport
+import spray.json.DefaultJsonProtocol._
 
 import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
 
-trait TrialServiceSupport extends LazyLogging {
+trait TrialServiceSupport extends LazyLogging with SprayJsonSupport {
 
-  val trialDAO:TrialDAO
+  val trialDAO: TrialDAO
+  val samDao: SamDAO
+  val thurloeDao: ThurloeDAO
+  implicit protected val executionContext: ExecutionContext
 
   private val dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
   private val spreadsheetFormat = new SimpleDateFormat("EEE, d MMM yyyy HH:mm:ss ")
@@ -178,6 +187,96 @@ trait TrialServiceSupport extends LazyLogging {
       currentStatus.copy(userId = userInfo.userSubjectId, state = Some(Enabled), userAgreed = false, enabledDate = Instant.now, billingProjectName = Some(trialProject.name.value))
     } else {
       currentStatus.copy(state = Some(Enabled))
+    }
+  }
+
+  def enableUserPostProcessing(updateStatus: Attempt, prevStatus: UserTrialStatus, newStatus: UserTrialStatus): Unit = {
+    if (updateStatus != StatusUpdate.Success && newStatus.billingProjectName.isDefined) {
+      logger.warn(
+        s"[trialaudit] The user '${newStatus.userId}' failed to be enabled, releasing the billing project '${newStatus.billingProjectName.get}' back into the available pool.")
+      trialDAO.releaseProjectRecord(RawlsBillingProjectName(newStatus.billingProjectName.get))
+    }
+  }
+
+  def executeStateTransitions(managerInfo: UserInfo, userEmails: Seq[String],
+                                      statusTransition: (WorkbenchUserInfo, UserTrialStatus) => UserTrialStatus,
+                                      transitionPostProcessing: (Attempt, UserTrialStatus, UserTrialStatus) => Unit): Future[PerRequestMessage] = {
+
+    def checkAndUpdateState(userInfo: WorkbenchUserInfo, userTrialStatus: UserTrialStatus, newStatus: UserTrialStatus): Future[String] = {
+      Try(requiresStateTransition(userTrialStatus, newStatus)) match {
+        case Success(isRequired) =>
+          if (isRequired) {
+            updateTrialStatus(managerInfo, userInfo, newStatus) map { stateResponse =>
+              transitionPostProcessing(stateResponse, userTrialStatus, newStatus)
+              StatusUpdate.toName(stateResponse)
+            }
+          } else {
+            logger.info(s"The user '${userInfo.userEmail}' is already in the trial state of '${newStatus.state.getOrElse("")}'. No further action will be taken.")
+            Future.successful(StatusUpdate.toName(StatusUpdate.NoChangeRequired))
+          }
+        case Failure(t: Throwable) => Future.successful(throwableToStatus(t))
+      }
+    }
+
+    val userTransitions = userEmails.map { userEmail =>
+      val status = Try(for {
+        regInfo <- samDao.adminGetUserByEmail(RawlsUserEmail(userEmail))
+        subId = regInfo.userInfo.userSubjectId
+        userInfo = WorkbenchUserInfo(subId, userEmail)
+        userTrialStatus <- thurloeDao.getTrialStatus(subId, managerInfo)
+        newStatus = statusTransition(userInfo, userTrialStatus)
+        result <- checkAndUpdateState(userInfo, userTrialStatus, newStatus)
+      } yield result) match {
+        case Success(s) => s
+        case Failure(t: Throwable) =>
+          Future.successful(throwableToStatus(t))
+      }
+      status map { finalStatus: String => (finalStatus, userEmail) } recover {
+        case ex: Exception =>
+          (StatusUpdate.toName(StatusUpdate.ServerError(ex.getMessage)), userEmail)
+      }
+    }
+
+    Future.sequence(userTransitions) map { results =>
+      val sorted: Map[String, Seq[String]] = results.groupBy(_._1).map { case (k, v) => (k, v.map(_._2)) }
+      RequestComplete(sorted)
+    }
+  }
+
+  def updateTrialStatus(managerInfo: UserInfo,
+                                userInfo: WorkbenchUserInfo,
+                                updatedTrialStatus: UserTrialStatus): Future[StatusUpdate.Attempt] = {
+    thurloeDao.saveTrialStatus(userInfo.userSubjectId, managerInfo, updatedTrialStatus) map {
+      case Success(_) =>
+        logger.info(s"[trialaudit] updated user ${userInfo.userEmail} (${userInfo.userSubjectId}) to state ${updatedTrialStatus.state.getOrElse("")}")
+        StatusUpdate.Success
+      case Failure(ex) => StatusUpdate.ServerError(ex.getMessage)
+    }
+  }
+
+  def requiresStateTransition(currentState: UserTrialStatus, newState: UserTrialStatus): Boolean = {
+    // this check needs to happen first for idempotent behavior
+    if (currentState.state == newState.state)
+      false
+    else {
+      // make sure the state transition is valid
+      if (newState.state.isEmpty || !newState.state.get.isAllowedFrom(currentState.state))
+        throw new FireCloudException(s"Cannot transition from ${currentState.state} to $newState.state.get")
+      true // indicates transition is required
+    }
+  }
+
+
+  def throwableToStatus(t: Throwable): String = {
+    t match {
+      case exr: FireCloudExceptionWithErrorReport =>
+        if (exr.errorReport.statusCode.contains(StatusCodes.NotFound))
+          StatusUpdate.toName(StatusUpdate.Failure("User not registered"))
+        else StatusUpdate.toName(StatusUpdate.Failure(exr.errorReport.message))
+      case ex: FireCloudException =>
+        StatusUpdate.toName(StatusUpdate.Failure(ex.getMessage))
+      case _ =>
+        StatusUpdate.toName(StatusUpdate.ServerError(t.getMessage))
     }
   }
